+++ app/models/maths_and_physics.rb
@@ -1,0 +1,9 @@
+module MathsAndPhysics
+  def self.start_page_url
+    "/maths-and-physics/start" # Temporary start page during private beta
+  end
+
+  def self.routing_name
+    "maths-and-physics"
+  end
+end