+++ app/models/maths_and_physics/slug_sequence.rb
@@ -1,0 +1,25 @@
+module MathsAndPhysics
+  # Determines the slugs that make up the claim process for a Maths & Physics
+  # claim. Based on the existing answers on the claim, the sequence of slugs
+  # will change. For example, if the claimant has said they are not paying off a
+  # student loan, the questions to determine their loan plan will not be part of
+  # the sequence.
+  #
+  # Note that the sequence is recalculated on each call to `slugs` so that it
+  # accounts for any changes that may have been made to the claim and always
+  # reflects the sequence based on the claim's current state.
+  class SlugSequence
+    SLUGS = [].freeze
+
+    attr_reader :claim
+
+    def initialize(claim)
+      @claim = claim
+    end
+
+    def slugs
+      SLUGS.dup.tap do |sequence|
+      end
+    end
+  end
+end